add workout details component

+++ server/controllers/workout-controller.js
@@ -9,13 +9,18 @@ const getAllWorkouts = async (req, res) => {
 // GET SINGLE
 const getSignleWorkout = async (req, res) => {
   const { id } = req.params;
-  try {
-    const workout = await workoutModel.findById(id);
-    res.status(200).json(workout);
-  } catch (err) {
-    console.error(err);
-    res.status(400).send("Something went wrong while fetching single workout.");
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({ error: "No such workout" });
   }
+
+  const workout = await workoutModel.findById(id);
+
+  if (!workout) {
+    return res.status(404).json({ error: "No such workout" });
+  }
+
+  res.status(200).json(workout);
 };
 
 // CREATE NEW
@@ -23,13 +28,11 @@ const createWorkout = async (req, res) => {
   const { title, reps, weight } = req.body;
 
   try {
-    if (!title || !reps || !weight) {
-      res.status(400).send("Please provide all fileds.");
-    }
-
+    // if (!title || !reps || !weight) {
+    //   return res.status(400).json({ error: "Please provide all fields." });
+    // }
     const workout = await workoutModel.create({ title, reps, weight });
-
-    res.status(200).json({ workout: workout });
+    res.status(200).json(workout);
   } catch (error) {
     res.status(400).json({ error: error.message });
   }
@@ -39,24 +42,37 @@ const createWorkout = async (req, res) => {
 const deleteWorkout = async (req, res) => {
   const { id } = req.params;
 
-  try {
-    await workoutModel.findOneAndDelete({ _id: id });
-    res.status(200).json({ message: "Workout deleted succesfully." });
-  } catch (error) {
-    res.status(400).json({ error: error.message });
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).json({ error: "No such workout" });
   }
+
+  const deletedWorkout = await workoutModel.findOneAndDelete({ _id: id });
+
+  if (!deletedWorkout) {
+    return res.status(404).json({ error: "Workout not found." });
+  }
+
+  res.status(200).json({ message: "Workout deleted successfully." });
 };
 
 // UPDATE
 const updateWorkout = async (req, res) => {
   const { id } = req.params;
 
-  try {
-    await workoutModel.findOneAndUpdate({ _id: id }, { ...req.body });
-    res.status(200).json({ message: "Workout updated succesfully." });
-  } catch (error) {
-    res.status(400).json({ error: error.message });
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).json({ error: "No such workout" });
   }
+
+  const updatedWorkout = await workoutModel.findOneAndUpdate(
+    { _id: id },
+    { ...req.body }
+  );
+
+  if (!updatedWorkout) {
+    return res.status(404).json({ error: "Workout not found." });
+  }
+
+  res.status(200).json({ message: "Workout updated successfully." });
 };
 
 module.exports = {
